Factor out test utility code and rename streaming to server
